requirements.txt
aws-cdk-lib==2.8.0
constructs>=10.0.0,<11.0.0
git+https://github.com/Jlrine2/aws-cdk-python-constructs.git